Change the algorithm for determining shapes from drawing bitmaps to calculating the crossing number.
Use the crossing number method to determine whether the rightmost point of an inner contour is
inside an outer contour. Test this for all outer contours.

This is more reliable than the previous method of drawing all outer contours and looking for the
nearest outer contour point to the right of the rightmost point of an inner contour in the
bitmap.

See http://geomalgorithms.com/a03-_inclusion.html