Use TableLayoutPanel to fix controls that are anchored top and bottom.
This applies to FontInfoForm and WindowsForm. It will be useful if you change the font.
Otherwise, controls that are anchored at the top and bottom -- like the license text box
in WindowsForm and window list box in WindowsForm -- are stretched outside of the form.

The solution is to put such controls docked in a cell of a TableLayoutPanel that's anchored
top and bottom or docked. The size of the row that contains the problematic control needs to be
set to 100%.